requirements.txt
wheel
Flask
Jinja2
Flask-Login
Flask-SQLAlchemy
psycopg2-binary